split percpu vs cpuinfo
