+++ include/kernel/infrastructure/i686/asm/cpuid.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Philippe Aubertin.
+ * Copyright (C) 2019-2024 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
@@ -29,47 +29,8 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_CPU_H
-#define JINUE_KERNEL_INFRASTRUCTURE_I686_CPU_H
-
-#include <kernel/infrastructure/i686/types.h>
-
-
-#define MSR_IA32_SYSENTER_CS        0x174
-
-#define MSR_IA32_SYSENTER_ESP       0x175
-
-#define MSR_IA32_SYSENTER_EIP       0x176
-
-#define MSR_EFER                    0xC0000080
-
-#define MSR_STAR                    0xC0000081
-
-
-#define MSR_FLAG_EFER_SCE           (1<<0)
-
-#define MSR_FLAG_EFER_NXE           (1<<11)
-
-
-#define CPU_FEATURE_CPUID           (1<<0)
-
-#define CPU_FEATURE_SYSENTER        (1<<1)
-
-#define CPU_FEATURE_SYSCALL         (1<<2)
-
-#define CPU_FEATURE_LOCAL_APIC      (1<<3)
-
-#define CPU_FEATURE_PAE             (1<<4)
-
-#define CPU_FEATURE_PGE             (1<<5)
-
-#define CPU_FEATURE_PSE             (1<<6)
-
-#define CPU_FEATURE_NOEXEC          (1<<7)
-
-
-#define CPU_EFLAGS_ID               (1<<21)
-
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_ASM_CPUID_H
+#define JINUE_KERNEL_INFRASTRUCTURE_I686_ASM_CPUID_H
 
 #define CPUID_FEATURE_FPU           (1<<0)
 
@@ -92,41 +53,12 @@
 
 #define CPUID_EXT_FEATURE_SYSCALL   (1<<11)
 
+#define CPUID_VENDOR_AMD_DW0        0x68747541    /* Auth */
+#define CPUID_VENDOR_AMD_DW1        0x69746e65    /* enti */
+#define CPUID_VENDOR_AMD_DW2        0x444d4163    /* cAMD */
 
-#define CPU_VENDOR_GENERIC          0
-
-#define CPU_VENDOR_AMD              1
-
-#define CPU_VENDOR_INTEL            2
-
-
-#define CPU_VENDOR_AMD_DW0          0x68747541    /* Auth */
-#define CPU_VENDOR_AMD_DW1          0x69746e65    /* enti */
-#define CPU_VENDOR_AMD_DW2          0x444d4163    /* cAMD */
-
-#define CPU_VENDOR_INTEL_DW0        0x756e6547    /* Genu */
-#define CPU_VENDOR_INTEL_DW1        0x49656e69    /* ineI */
-#define CPU_VENDOR_INTEL_DW2        0x6c65746e    /* ntel */
-
-typedef struct {
-    unsigned int     maxphyaddr;
-    unsigned int     dcache_alignment;
-    uint32_t         features;
-    int              vendor;
-    int              family;
-    int              model;
-    int              stepping;
-} cpu_info_t;
-
-extern cpu_info_t cpu_info;
-
-static inline bool cpu_has_feature(uint32_t mask) {
-    return (cpu_info.features & mask) == mask;
-}
-
-void cpu_init_data(cpu_data_t *data);
-
-void cpu_detect_features(void);
-
+#define CPUID_VENDOR_INTEL_DW0      0x756e6547    /* Genu */
+#define CPUID_VENDOR_INTEL_DW1      0x49656e69    /* ineI */
+#define CPUID_VENDOR_INTEL_DW2      0x6c65746e    /* ntel */
 
 #endif
