+++ include/kernel/infrastructure/i686/asm/msr.h
@@ -1,22 +1,22 @@
 /*
- * Copyright (C) 2024 Philippe Aubertin.
+ * Copyright (C) 2019-2024 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,15 +29,22 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <kernel/application/syscalls.h>
-#include <kernel/domain/entities/thread.h>
-#include <kernel/machine/thread.h>
-#include <kernel/machine/tls.h>
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_ASM_MSR_H
+#define JINUE_KERNEL_INFRASTRUCTURE_I686_ASM_MSR_H
 
-void set_thread_local(void *addr, size_t size) {
-    thread_t *thread = get_current_thread();
+#define MSR_IA32_SYSENTER_CS        0x174
 
-    set_thread_local_storage(thread, addr, size);
-    
-    machine_set_tls(thread);
-}
+#define MSR_IA32_SYSENTER_ESP       0x175
+
+#define MSR_IA32_SYSENTER_EIP       0x176
+
+#define MSR_EFER                    0xC0000080
+
+#define MSR_STAR                    0xC0000081
+
+
+#define MSR_FLAG_EFER_SCE           (1<<0)
+
+#define MSR_FLAG_EFER_NXE           (1<<11)
+
+#endif
