+++ include/kernel/infrastructure/i686/percpu.h
@@ -1,22 +1,22 @@
 /*
- * Copyright (C) 2024 Philippe Aubertin.
+ * Copyright (C) 2019-2024 Philippe Aubertin.
  * All rights reserved.
 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,15 +29,28 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <kernel/application/syscalls.h>
-#include <kernel/domain/entities/thread.h>
-#include <kernel/machine/thread.h>
-#include <kernel/machine/tls.h>
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_PERCPU_H
+#define JINUE_KERNEL_INFRASTRUCTURE_I686_PERCPU_H
 
-void set_thread_local(void *addr, size_t size) {
-    thread_t *thread = get_current_thread();
+#include <kernel/infrastructure/i686/types.h>
+#include <kernel/infrastructure/i686/x86.h>
 
-    set_thread_local_storage(thread, addr, size);
-    
-    machine_set_tls(thread);
+#define PERCPU_DATA_ALIGNMENT 256
+
+static inline percpu_t *get_percpu_data(void) {
+    /* CPU data structure is at address zero within the per-cpu data segment */
+    percpu_t *const zero = (percpu_t *)0;
+    return (percpu_t *)get_gs_ptr( (uint32_t *)&zero->self );
 }
+
+static inline tss_t *get_tss(void) {
+    return &get_percpu_data()->tss;
+}
+
+static inline addr_space_t *get_current_addr_space(void) {
+    return get_percpu_data()->current_addr_space;
+}
+
+void init_percpu_data(percpu_t *data);
+
+#endif
