+++ include/kernel/machine/tls.h
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,15 +29,11 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <kernel/application/syscalls.h>
-#include <kernel/domain/entities/thread.h>
-#include <kernel/machine/thread.h>
-#include <kernel/machine/tls.h>
+#ifndef JINUE_KERNEL_MACHINE_TLS_H
+#define JINUE_KERNEL_MACHINE_TLS_H
 
-void set_thread_local(void *addr, size_t size) {
-    thread_t *thread = get_current_thread();
+#include <kernel/types.h>
 
-    set_thread_local_storage(thread, addr, size);
-    
-    machine_set_tls(thread);
-}
+void machine_set_tls(const thread_t *thread);
+
+#endif
