+++ kernel/infrastructure/i686/init.c
@@ -34,6 +34,7 @@
 #include <kernel/domain/services/cmdline.h>
 #include <kernel/domain/services/logging.h>
 #include <kernel/domain/services/panic.h>
+#include <kernel/infrastructure/i686/asm/msr.h>
 #include <kernel/infrastructure/i686/drivers/pic8259.h>
 #include <kernel/infrastructure/i686/drivers/uart16550a.h>
 #include <kernel/infrastructure/i686/drivers/vga.h>
@@ -41,10 +42,10 @@
 #include <kernel/infrastructure/i686/pmap/pae.h>
 #include <kernel/infrastructure/i686/pmap/pmap.h>
 #include <kernel/infrastructure/i686/boot_alloc.h>
-#include <kernel/infrastructure/i686/cpu.h>
-#include <kernel/infrastructure/i686/cpu_data.h>
+#include <kernel/infrastructure/i686/cpuinfo.h>
 #include <kernel/infrastructure/i686/descriptors.h>
 #include <kernel/infrastructure/i686/memory.h>
+#include <kernel/infrastructure/i686/percpu.h>
 #include <kernel/infrastructure/i686/x86.h>
 #include <kernel/infrastructure/elf.h>
 #include <kernel/interface/i686/boot.h>
@@ -97,7 +98,7 @@ static bool maybe_enable_pae(
 
     bool use_pae;
 
-    if(cpu_has_feature(CPU_FEATURE_PAE)) {
+    if(cpu_has_feature(CPUINFO_FEATURE_PAE)) {
         use_pae = (config->machine.pae != CONFIG_PAE_DISABLE);
     }
     else {
@@ -145,7 +146,7 @@ static void init_idt(void) {
     }
 }
 
-static void load_selectors(cpu_data_t *cpu_data, boot_alloc_t *boot_alloc) {
+static void load_selectors(percpu_t *cpu_data, boot_alloc_t *boot_alloc) {
     /* Pseudo-descriptor allocation is temporary for the duration of this
      * function only. Remember heap pointer on entry so we can free the
      * pseudo-allocator when we are done. */
@@ -203,7 +204,7 @@ static void remap_text_video_memory(boot_alloc_t *boot_alloc) {
 }
 
 static void enable_global_pages(void) {
-    if(cpu_has_feature(CPU_FEATURE_PGE)) {
+    if(cpu_has_feature(CPUINFO_FEATURE_PGE)) {
         set_cr4(get_cr4() | X86_CR4_PGE);
     }
 }
@@ -219,7 +220,7 @@ static void initialize_page_allocator(boot_alloc_t *boot_alloc) {
 }
 
 static void select_syscall_implementation(void) {
-    if(cpu_has_feature(CPU_FEATURE_SYSCALL)) {
+    if(cpu_has_feature(CPUINFO_FEATURE_SYSCALL)) {
         uint64_t msrval;
 
         syscall_implementation = JINUE_I686_HOWSYSCALL_FAST_AMD;
@@ -234,7 +235,7 @@ static void select_syscall_implementation(void) {
 
         wrmsr(MSR_STAR, msrval);
     }
-    else if(cpu_has_feature(CPU_FEATURE_SYSENTER)) {
+    else if(cpu_has_feature(CPUINFO_FEATURE_SYSENTER)) {
         syscall_implementation = JINUE_I686_HOWSYSCALL_FAST_INTEL;
 
         wrmsr(MSR_IA32_SYSENTER_CS,  SEG_SELECTOR(GDT_KERNEL_CODE, RPL_KERNEL));
@@ -310,7 +311,7 @@ void machine_init(const config_t *config) {
 
     const bootinfo_t *bootinfo = get_bootinfo();
 
-    cpu_detect_features();
+    detect_cpu_features();
 
     check_data_segment(bootinfo);
 
@@ -338,11 +339,11 @@ void machine_init(const config_t *config) {
      * 
      * We need to ensure that the Task State Segment (TSS) contained in this
      * memory block does not cross a page boundary. */
-    assert(sizeof(cpu_data_t) < CPU_DATA_ALIGNMENT);
-    cpu_data_t *cpu_data = boot_heap_alloc(&boot_alloc, cpu_data_t, CPU_DATA_ALIGNMENT);
+    assert(sizeof(percpu_t) < PERCPU_DATA_ALIGNMENT);
+    percpu_t *cpu_data = boot_heap_alloc(&boot_alloc, percpu_t, PERCPU_DATA_ALIGNMENT);
     
     /* initialize per-CPU data */
-    cpu_init_data(cpu_data);
+    init_percpu_data(cpu_data);
     
     /* Initialize interrupt descriptor table (IDT) */
     init_idt();
